I'm interested in weatherization program of public housing
I have lived in public housing for 8 years. My electric bill is almost as much as my rent. I know for sure there is not adequate insulation in the Attic. And my landlord and maintenance man refuse to address the problem.
Simple caulking of the windows inside and out and proper insulation would save the tenants a lot of cost for heating and cooling.